refactor(engine): move revdeps to engine as a streaming Engine::revdeps
Relocate the reverse-dependency search out of the inspect command into
the engine, alongside query, as `Engine::revdeps`. It now returns a
`Stream<Item = Result<Addr>>` — candidates are consumed straight off the
query stream, resolved concurrently (bounded by 2×parallelism), and users
are yielded as found. Nothing is collected in bulk; the command prints
each user incrementally via BufferedStdout.

Output order follows completion order rather than a stable sort; callers
that need determinism collect and sort (the engine tests do).
